Sensor - v4 - Refactor Collection
(1) Add Collection class; CollectIP and CollectHTTPRequest
(2) Comment out legacy methods in Program.cs
(3) Refactor "Load" operation
(4) Coment out legacy "Load" method
(5) Clean out legacy pervious commented code
(6) Add Console loggs for timer status
